Add logging functionality to FoliCon Modules
Added NLog logic for debugging and tracing into IGDBClass and Extensions in FoliCon Modules. The loggers provide necessary debug information such as search queries, the ID being used for a search or when fetching videos, and warn if no poster is found for a particular result. Furthermore, the changes trace both the initiation and completion of the GetBitmap from HttpResponseMessage method.