Handle unauthorized file access, improve download failure handling
This commit handles any UnauthorizedAccessException during the manipulation of files i.e., during the deletion of file icons and saving of media info. The changes now ensure that the user is informed if they do not have the required permissions to access/alter the files.

Additionally, the downloading of images from URLs has been improved. Now, in case of failures it will retry up to 2 times and eventually, if all attempts fail, it will log an error message. This should provide more robust handling of network issues or transient problems with the remote server hosting the images. (Close #120)