Add Error Reporting Feature in FoliCon
The commit introduces a feature for error reporting in the application FoliCon. A new boolean variable _enableErrorReporting is added in the MainWindowViewModel.cs file, which is by default set to false. The Sentry configuration is added or removed based on the boolean's value. This option is provided to the user in the Main Window, allowing users to enable or disable error reporting as per their preference. Necessary changes are made to update the user interface language files. This improves the debugging process and provides user control over error reporting.