Update HTMLBox view and style
Refactor HTMLBox.xaml related code to improve the loading experience. A new `System.Windows.Navigation` namespace was added and some changes were made to the style and structure of the layout.

The browser visibility is initially set to hidden and only toggles to visible once the loading is complete. This is done to prevent a flash of styled content.

The HTML skeleton was also updated to make it more robust and to prevent overflow. Removed static background colors and replaced them with dynamic ones for better theme management.

Incorporated these changes also in `SearchResult.xaml` and `HtmlBox.xaml` for consistent UX across the application.

The Update aims for a smoother user experience, especially when the application theme is toggled and while the browser component is loading the content.