Add TextToVisibility converters to handle unavailability of videos
Two new classes 'TextToVisibilityConverter' and 'InvertedTextToVisibilityConverter' have been added to the 'Convertor' module that convert the input text to visibility based on whether the related video is available or not. These are needed to enhance user experience by managing visibility of UI components corresponding to the video availability. SearchResult.xaml has been modified to use these converters. It displays a placeholder image when a video is not available and otherwise allows the HTML text of the video to be loaded.