Remove TMDB API service and data transformation logic from TMDB helper file
The TMDB file was becoming cluttered with a mix of data transformation logic and service level API interactions. For better maintainability and separation of concerns, moved TMDB API service logic into a new file `TMDBService.cs` and data transformation logic into a new file `TmdbDataTransformer.cs`. This change does not affect the functional behavior of the application, it refactors the codebase for better modularity and readability.