Implement caching for DeviantArt API access token
The DeviantArt module was checking with server for token validation every time it needed one. This resulted in network calls even when the existing one was valid. To improve efficiency, caching has been implemented using Microsoft's MemoryCache. This allows the module to use a valid token from the cache instead of generating a new one. URL constructions in the DArt class are also refactored into dedicated methods for better readability and maintenance.