Add GC suppression and remove default customRating
In MainWindowViewModel.cs, added "GC.SuppressFinalize(this);" to prevent the garbage collector from finalizing the instance.
In SearchResultViewModel.cs, removed the default value for the "_customRating" field. The GC was suppressed to improve efficiency and "_customRating" was updated as the default value was not required and could lead to misinterpretation.