Add method to check if folder is a series folder
A new function named 'IsSeriesFolder' has been added to FileUtils.cs. This function checks if a given folder path corresponds to a series folder by looking for a pattern match with the term 'season' followed by a number in the subfolders' names. Any errors that occur during this process are logged.