Replace WebBrowser with WebView2 in HtmlBox view
This commit introduces the replacement of WebBrowser with WebView2 in the HtmlBox view of the FoliCon application. The WebView2 control provides more capabilities and better performance compared to the older WebBrowser control. Accordingly, the code is modified to use the functionality of WebView2, which includes the 'NavigateToString' method being called on the 'CoreWebView2' object instead of the outdated 'Browser' object. The 'InitializeAsync' method has also been added to ensure the WebView2 control is correctly initialized.