Add DeviationId to DArtImageList and create ExtractManually method
DArtImageList model has been expanded to include the DeviationId property. This required the constructor and relevant bindings to be adjusted accordingly. Further, an ExtractManually method has been added to ProSearchResultViewModel, allowing a more direct extraction from a given Deviation ID. Also, the model binding in ProSearchResult.xaml was modified for the tag to incorporate the newly added DeviationId property.