Feature: Feature to extract Deviations from DeviantArt, so that in professional mode, use can pick from images, that are in an compressed archive (#203)
* Add Dummy image context menu

* Refactor: make serializationSetting a class field and use string interpolation for placebo url

* Refactor directory permission checks and add a method to check if the file is a compressed archive by filename
 extension

The directory permission checks were refactored to use .NET's built-in System.Security classes instead of a manual try-catch approach. This provides a more accurate and robust permission checking ability. Additionally, functionality was added to detect if a file is a compressed archive by checking its extension against a predefined list of compression extensions. Other assorted methods related to directory handling and filesystem utilities were also added.

* feature: Added download functionality to DeviantArt module

This change includes functionality to download DeviantArt deviations, process compressed files and save them to a local path. This functionality needed a new library, 'SharpCompress,' which has been added to the project. In addition, new models have been added to handle API responses, and extensions were added for archive stream processing.

* Add DeviationId to DArtImageList and create ExtractManually method

DArtImageList model has been expanded to include the DeviationId property. This required the constructor and relevant bindings to be adjusted accordingly. Further, an ExtractManually method has been added to ProSearchResultViewModel, allowing a more direct extraction from a given Deviation ID. Also, the model binding in ProSearchResult.xaml was modified for the tag to incorporate the newly added DeviationId property.

* feature: Add ManualExplorer view and related ViewModel

Added ManualExplorer view to the FoliCon application, enabling manual extraction of images. The ManualExplorer view is designed to display images from a specific directory and allows users to interact with them. Additionally, a ViewModel was created for this view to handle its associated functionalities. The DialogServiceExtensions class is also updated to provide dialog services for the new view.

* fix: change target directory path identifier

Changed the identifier for the target directory path from the dArtDownloadResponse.Filename to the deviationId for fixing the potential long path issue.

* Complete pickMethod for ManualExtraction

A check has been added to differentiate between local and remote image paths when downloading images as files are local in case of ManualExtraction. The "FrameworkElement_OnUnloaded" event has been added to ensure image sources are properly released and no lock is present on them.

* Refactor: Cleanup tempDeviation directory at FoliCon close event

* Chore: Fix Codacy issues

This commit includes several crucial refactorings across multiple files in the project. Redundant code has been replaced by more efficient functions, such as the replacement of the entire GetFileSizeHumanReadable() method with the ConvertHelper.ToFileSize() function. The DArtImageList now uses base constructors for improved cleanliness. Also, the 'if' statement formatting of ProSearchResultViewModel.cs, and FileUtils.cs are streamlined, and a more descriptive method in StreamExtension.cs makes the flow of control more visible.

* remove redundant code

* Add translation strings for the manual extraction feature

* Add comments to DeviantArt and StreamExtension methods

Added docstrings to the `Download` method in the `DArt.cs` and `ExtractPngAndIcoToDirectory` method in `StreamExtension.cs` to provide more clarity about their usage and output.