Add Cancel option and progress for manual extraction (#205)
* feature: Implement extraction progress tracking

Updated ManualExplorer_views and ManualExplorerViewModel to include progress tracking for file downloads. This involves adding new properties to the model and bindings in the UI. Advanced Handling of extraction of files is also added in DArt.cs. The error scenario while deleting temporary directories is also considered.

* Fix: Update ExtractionProgress.cs to use Bendable base to notify the UI of the value change, as the progress bar wouldn't update and format progress bar text

* Refactor progress model and add cancel functionality

ExtractionProgress class has been renamed to ProgressInfo and extended with a text property to provide more detailed progress information. Cancel functionality has been added to the manual extraction process, allowing the user to terminate the process prematurely. Fetching and extraction processes have been updated to be cancellable. In addition, the progress display in the UI has been updated to provide more detailed information.

* fix: Add ImageCacheConverter to avoid lock on files when loading in ManualExplorer

* Add localized strings for downloading and extracting status

This commit encompasses the addition of localized string definitions for 'Downloading' and 'Extracting' scenarios.

* Fix codacy style issues, and CTS being disposed of before the busyIndicator vanishes