Refactor: Adjust commands to handle DArtImageList object directly
The PickCommand and OpenImageCommand in the ProSearchResultViewModel have been refactored to take a DArtImageList object directly, instead of an object. This means we can handle the "MustWatch" property in the commands, allowing us to manage scenarios where the image needs to be watched to be downloaded. Additionally, a new property 'PremiumFolderData' has been added to 'DArtBrowseResult.cs', allowing tracking of folders with premium accessibility. Corresponding adjustments were made on the UI side in ProSearchResult.xaml. The 'DArtImageList' constructor is also updated to accept a new parameter - 'mustWatch'.