Fix: Folder count now takes into account subfolders and also fixes the issue of incorrect processed count
Properly update the total folder count by not only considering parent folders but also adding the count of subfolders found during the search process. It also resolves an issue where a potentially incorrect folder processed count was returned when the poster window was not shown and 'Skip Ambiguous' was checked.