Fix: Incorrect total folder count when subfolder processing is enabled and Incorrect processed count in a case where skipAmbigous was true and there were multiple results (#209)
Properly update the total folder count by not only considering parent folders but also adding the count of subfolders found during the search process. It also resolves an issue where a potentially incorrect folder processed count was returned when the poster window was not shown and 'Skip Ambiguous' was checked.